+++ twitchio/client.py
@@ -308,16 +308,16 @@ class Client:
 
             return Channel(name=name, websocket=self._connection)
 
-    async def leave_channels(self, channels: Union[List[str], Tuple[str]]):
+    async def part_channels(self, channels: Union[List[str], Tuple[str]]):
         """|coro|
 
 
-        Leave the specified channels.
+        Part from the specified channels.
 
         Parameters
         ------------
         channels: Union[List[str], Tuple[str]]
-            The channels in either a list or tuple form to leave.
+            The channels in either a list or tuple form to part from.
         """
         await self._connection.part_channels(*channels)